fix usecase-roles types
