--- examples/usecase-roles/types.ts
@@ -1,22 +1,0 @@
-export type Session = {
-  itemId: string;
-  listKey: string;
-  data: {
-    name: string;
-    role?: {
-      id: string;
-      name: string;
-      canCreateTodos: boolean;
-      canManageAllTodos: boolean;
-      canSeeOtherPeople: boolean;
-      canEditOtherPeople: boolean;
-      canManagePeople: boolean;
-      canManageRoles: boolean;
-    };
-  };
-};
-
-export type ListAccessArgs = {
-  itemId?: string;
-  session?: Session;
-};
